# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SkillScope – AI-Powered Developer Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /resumes:
    get:
      operationId: getResumes
      tags: [resumes]
      summary: List all resumes
      responses:
        "200":
          description: Resumes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resume"
    post:
      operationId: createResume
      tags: [resumes]
      summary: Create a new resume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeInput"
      responses:
        "201":
          description: Resume created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
  /resumes/{id}:
    get:
      operationId: getResume
      tags: [resumes]
      summary: Get a specific resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
        "404":
          description: Not found
    patch:
      operationId: updateResume
      tags: [resumes]
      summary: Update a resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeUpdate"
      responses:
        "200":
          description: Updated resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
    delete:
      operationId: deleteResume
      tags: [resumes]
      summary: Delete a resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /resumes/{id}/generate:
    post:
      operationId: generateResume
      tags: [resumes]
      summary: AI-generate resume content from connected profile data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generated resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"

  # ─── Interview Readiness ─────────────────────────────────────────────────────
...
  /recruiter/search:
    get:
      operationId: searchDevelopers
      tags: [recruiter]
      summary: Search developers by skills, scores, and location
      parameters:
        - name: skills
          in: query
          schema:
            type: string
          description: Comma-separated skill names
        - name: minGithubScore
          in: query
          schema:
            type: integer
        - name: minCodingScore
          in: query
          schema:
            type: integer
        - name: location
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeveloperSearchResult"
  /recruiter/bookmarks:
    get:
      operationId: getRecruiterBookmarks
      tags: [recruiter]
      summary: Get bookmarked developers
      responses:
        "200":
          description: Bookmarks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecruiterBookmark"
    post:
      operationId: addRecruiterBookmark
      tags: [recruiter]
      summary: Bookmark a developer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecruiterBookmarkInput"
      responses:
        "201":
          description: Bookmarked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecruiterBookmark"
  /recruiter/bookmarks/{userId}:
    delete:
      operationId: removeRecruiterBookmark
      tags: [recruiter]
      summary: Remove a developer bookmark
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed

  # ─── Public Profiles ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Resume:
      type: object
      required: [id, title, template, resumeScore, atsScore, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        template:
          type: string
          enum: [fresher, experienced, product, startup, internship]
        content:
          type: ["object", "null"]
          description: Structured resume content as JSON
        resumeScore:
          type: integer
        atsScore:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    ResumeInput:
      type: object
      required: [title, template]
      properties:
        title:
          type: string
        template:
          type: string
          enum: [fresher, experienced, product, startup, internship]
    ResumeUpdate:
      type: object
      properties:
        title:
          type: string
        template:
          type: string
          enum: [fresher, experienced, product, startup, internship]
        content:
          type: object
          description: Structured resume content as JSON

    # ─── Interview ──────────────────────────────────────────────────────────────
...
    DeveloperSearchResult:
      type: object
      required: [total, developers]
      properties:
        total:
          type: integer
        developers:
          type: array
          items:
            $ref: "#/components/schemas/DeveloperCard"
    DeveloperCard:
      type: object
      required: [userId, name, overallScore, githubScore, codingScore, interviewScore, topSkills]
      properties:
        userId:
          type: integer
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        college:
          type: ["string", "null"]
        overallScore:
          type: integer
        githubScore:
          type: integer
        codingScore:
          type: integer
        interviewScore:
          type: integer
        resumeScore:
          type: integer
        topSkills:
          type: array
          items:
            type: string
        username:
          type: ["string", "null"]
    RecruiterBookmark:
      type: object
      required: [id, bookmarkedUserId, bookmarkedUser, createdAt]
      properties:
        id:
          type: integer
        bookmarkedUserId:
          type: integer
        bookmarkedUser:
          $ref: "#/components/schemas/DeveloperCard"
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [saved, shortlisted, contacted, rejected]
        createdAt:
          type: string
    RecruiterBookmarkInput:
      type: object
      required: [bookmarkedUserId]
      properties:
        bookmarkedUserId:
          type: integer
        notes:
          type: string
        status:
          type: string
          enum: [saved, shortlisted, contacted, rejected]

    # ─── Public ─────────────────────────────────────────────────────────────────
...